Рефакторинг. Объединил 2 класса в один.
